some eddition 06.02.2017

+++ routes/web.php
@@ -1,46 +1,50 @@
 <?php
 
 Route::group(['prefix' => 'admin'], function (){
-    Route::get('/', 'AdminController@index');
-    Route::put('update/{id}', 'AdminController@update');
+
+    Route::get('/', 'AdminController@index')->middleware('auth');
+    Route::put('update/{id}', 'AdminController@update')->middleware('auth');
 
     Route::get('login', 'AdminController@getLogin');
     Route::post('login', 'AdminController@postLogin');
     Route::post('logout', 'AdminController@postLogout');
 
-    Route::get('blog', 'BlogController@blogShow');
-    Route::get('post-add', 'BlogController@postAdd');
-    Route::post('post-store', 'BlogController@postStore');
-    Route::get('post-edit/{id}', 'BlogController@postEdit');
-    Route::put('post-update/{id}', 'BlogController@postUpdate');
-    Route::delete('post-delete/{id}', 'BlogController@postDelete');
+    Route::get('blog', 'BlogController@blogShow')->middleware('auth');
+    Route::get('post-add', 'BlogController@postAdd')->middleware('auth');
+    Route::post('post-store', 'BlogController@postStore')->middleware('auth');
+    Route::get('post-edit/{id}', 'BlogController@postEdit')->middleware('auth');
+    Route::put('post-update/{id}', 'BlogController@postUpdate')->middleware('auth');
+    Route::delete('post-delete/{id}', 'BlogController@postDelete')->middleware('auth');
+    Route::get('post/drafts', 'BlogController@postDrafts')->middleware('auth');
 
-    Route::get('categories', 'CategoryController@catShow');
-    Route::get('category-add', 'CategoryController@catAdd');
-    Route::post('category-store', 'CategoryController@catStore');
-    Route::get('category-edit/{id}', 'CategoryController@catEdit');
-    Route::put('category-update/{id}', 'CategoryController@catUpdate');
-    Route::delete('category-delete/{id}', 'CategoryController@catDelete');
+    Route::get('categories', 'CategoryController@catShow')->middleware('auth');
+    Route::get('category-add', 'CategoryController@catAdd')->middleware('auth');
+    Route::post('category-store', 'CategoryController@catStore')->middleware('auth');
+    Route::get('category-edit/{id}', 'CategoryController@catEdit')->middleware('auth');
+    Route::put('category-update/{id}', 'CategoryController@catUpdate')->middleware('auth');
+    Route::delete('category-delete/{id}', 'CategoryController@catDelete')->middleware('auth');
 
-    Route::get('shop', 'ProductController@prodShow');
-    Route::get('product-add', 'ProductController@prodAdd');
-    Route::post('product-store', 'ProductController@prodStore');
-    Route::get('product-edit/{id}', 'ProductController@prodEdit');
-    Route::put('product-update/{id}', 'ProductController@prodUpdate');
-    Route::delete('product-delete/{id}', 'ProductController@prodDelete');
+    Route::get('shop', 'ProductController@prodShow')->middleware('auth');
+    Route::get('product-add', 'ProductController@prodAdd')->middleware('auth');
+    Route::post('product-store', 'ProductController@prodStore')->middleware('auth');
+    Route::get('product-edit/{id}', 'ProductController@prodEdit')->middleware('auth');
+    Route::put('product-update/{id}', 'ProductController@prodUpdate')->middleware('auth');
+    Route::delete('product-delete/{id}', 'ProductController@prodDelete')->middleware('auth');
+    Route::get('product/store', 'ProductController@prodInStore');
 
-    Route::get('gallery', 'GalleryController@galShow');
-    Route::get('image-add', 'GalleryController@galAdd');
-    Route::post('image-store', 'GalleryController@galStore');
-    Route::delete('image-delete/{id}', 'GalleryController@galDelete');
+    Route::get('gallery', 'GalleryController@galShow')->middleware('auth');
+    Route::get('image-add', 'GalleryController@galAdd')->middleware('auth');
+    Route::post('image-store', 'GalleryController@galStore')->middleware('auth');
+    Route::delete('image-delete/{id}', 'GalleryController@galDelete')->middleware('auth');
 
-    Route::get('orders', 'OrderController@orderShow');
-    Route::put('order-done/{id}', 'OrderController@orderDone');
+    Route::get('orders', 'OrderController@orderShow')->middleware('auth');
+    Route::put('order-done/{id}', 'OrderController@orderDone')->middleware('auth');
+    Route::get('order/arch', 'OrderController@archiveShow')->middleware('auth');
 
-    Route::get('contacts', 'ContactController@contactShow');
+    Route::get('contacts', 'ContactController@contactShow')->middleware('auth');
 
-    Route::delete('img-delete/{id}', 'FileController@deleteImg');
-    //Auth::routes();
+    Route::delete('img-delete/{id}', 'FileController@deleteImg')->middleware('auth');
+
 });
 
 Route::get('/', 'PageController@index');
@@ -60,3 +64,9 @@ Route::post('shop/order-store', 'OrderController@indexOrderStore');
 
 Route::get('gallery', 'GalleryController@indexShow');
 
+
+Auth::routes();
+
+Auth::routes();
+
+Route::get('/home', 'HomeController@index');
